+++ CHANGELOG.md
@@ -30,6 +30,49 @@
 - [[`81ad77f4a5`](https://github.com/twreporter/twreporter-react/commit/81ad77f4a5)] - **feat**: article tracking section (Lucien)
 - [[`b1bbe8ec56`](https://github.com/twreporter/twreporter-react/commit/b1bbe8ec56)] - **feat**: article tracking card (Lucien)
 
+## 5.0.5-rc.2, 2024-06-23
+
+### Notable Changes
+
+- chore
+  - update @twreporter packages
+
+### Commits
+
+- [[`4e6bb7867d`](https://github.com/twreporter/twreporter-react/commit/4e6bb7867d)] - **chore**: update twreporter packages (Lucien)
+
+## 5.0.5-rc.1, 2024-06-20
+
+### Notable Changes
+
+- feat
+  - add `useWebpush` hook & normalize Banner
+- chore
+  - update @twreporter packages
+
+### Commits
+
+- [[`14ce5f8054`](https://github.com/twreporter/twreporter-react/commit/14ce5f8054)] - **chore**: update @twreporter packages (Aylie Chou)
+- [[`5e41e26d7d`](https://github.com/twreporter/twreporter-react/commit/5e41e26d7d)] - **fix**: update import path for article-banner (Aylie Chou)
+- [[`b7fbdd3e4c`](https://github.com/twreporter/twreporter-react/commit/b7fbdd3e4c)] - **feat**: add `useWebpush` hook & normalize Banner (Aylie Chou)
+
+## 5.0.5-rc.0, 2024-06-19
+
+### Notable Changes
+
+- feat
+  - use react-redux hook
+- chore
+  - update @twreporter packages
+  - update to latest redux
+
+### Commits
+
+- [[`d3245ff100`](https://github.com/twreporter/twreporter-react/commit/d3245ff100)] - **chore**: update twreporter packages (Lucien)
+- [[`624ffb5cc6`](https://github.com/twreporter/twreporter-react/commit/624ffb5cc6)] - **feat**: use react-redux hook (Lucien)
+- [[`a09cbb24de`](https://github.com/twreporter/twreporter-react/commit/a09cbb24de)] - **fix**: remove yalc packages (Lucien)
+- [[`7ff15790d2`](https://github.com/twreporter/twreporter-react/commit/7ff15790d2)] - **chore**: update to latest redux (Lucien)
+
 ## 5.0.4, 2024-06-04 (Current)
 
 ### Notable Changes